+++ front/src/components/atoms/ShowModal.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ShowModal = () => {
+  
+  return(
+    <>
+    </>
+  )
+}
+
+export default ShowModal;
