+++ front/src/components/pages/TasksPage.jsx
@@ -1,10 +1,8 @@
 import React from "react";
 import { useGetTasks } from "../../stores/useTasks/useGetTasks";
-import { Link } from "react-router-dom";
 import Loading from "../atoms/Loading";
 import {
   Box,
-  Button,
   Table,
   Thead,
   Tbody,
@@ -12,8 +10,11 @@ import {
   Th,
   Td,
   TableContainer,
+  Text,
+  Center,
 } from '@chakra-ui/react'
 import DeleteButton from "../atoms/DeleteButton";
+import CreateForm from "../atoms/CreateForm";
 
 const Task = () => {
   const { data: tasks } = useGetTasks();
@@ -21,32 +22,43 @@ const Task = () => {
 
   return (
     <>
-      <Box maxH="500px" maxW="960px" mx="auto">
-        <Button colorScheme='teal' variant='solid'>
-          <Link to="/tasks/create">
-            新規作成
-          </Link>
-        </Button>
-        <TableContainer>
-          <Table variant='simple'>
-              <Thead>
-                <Tr>
-                  <Th>ID</Th>
-                  <Th>タイトル</Th>
-                  <Th>完了する（削除）</Th>
-                </Tr>
-              </Thead>
-              { tasks.map((item, index) => (
-                <Tbody key={item.id}>
-                  <Tr>
-                    <Td>{index + 1}</Td>
-                    <Td>{item.title}</Td>
-                    <Td><DeleteButton task={item} /></Td>
-                  </Tr>
-                </Tbody>
-              ))}
-          </Table>
-        </TableContainer>  
+      <Box maxH="100vh" maxW="960px" mx="auto">
+        <Box mb={5} >
+          <CreateForm />
+              {tasks.length !== 0 
+              ? 
+              (
+                <>
+                <TableContainer>
+                  <Table variant='simple'>
+                      <Thead>
+                        <Tr>
+                          <Th>ID</Th>
+                          <Th w={'200px'}>タイトル</Th>
+                          <Th>完了する（削除）</Th>
+                        </Tr>
+                      </Thead>
+                      { tasks.map((item, index) => (
+                        <Tbody key={item.id}>
+                          <Tr>
+                            <Td>{index + 1}</Td>
+                            <Td>{item.title}</Td>
+                            <Td><DeleteButton task={item} /></Td>
+                          </Tr>
+                        </Tbody>
+                      ))}
+                      </Table>
+                    </TableContainer>  
+              </>
+              ) : (
+                <>
+                <Center>
+                  
+                  <Text>タスクはありません！</Text>
+                </Center>
+              </>
+              )}
+        </Box>
       </Box>
     </>
   )
